Closure modification and new refactoring proofs
1. Modify the closure concept to be able to evaluate redefined functions.
For example a function like this:

let X = 5 in
  let X = fun() -> X in
    let X = fun() -> apply X() in
      apply X()

The pure reference is changed to a union type (either a concrete
environment is stored in the closures or a reference just like before).
This way it is enough to store function signatures in the closure
environment.

2. Modify the signature of the semantics inductive type to allow
currying. This way more powerful induction can be used.

3. Modify everything (examples, semantics, helper functions, proofs)
accordingly to 1 and 2.

4. Add new refactoring proofs (another form of the commutativity and the
use of a function introduced with let statement). Add hypotheses to the
old commutativity equivalence example and fully prove it.